requirements.txt
asgiref==3.4.1
Django==3.2.8
geographiclib==1.52
geopy==2.2.0
numpy==1.21.3
opencv-python==4.5.4.58
Pillow==8.4.0
psycopg2==2.9.1
pytz==2021.3
sqlparse==0.4.2
